organize utils in rpc
