+++ crates/client/rpc/src/madara_backend_client.rs
@@ -8,7 +8,7 @@ use sp_runtime::traits::{Block as BlockT, Header as HeaderT};
 use starknet_api::hash::StarkHash;
 
 use crate::errors::StarknetRpcApiError;
-use crate::utils::get_block_by_block_hash;
+use crate::get_block_by_block_hash;
 
 pub fn load_hash<C>(client: &C, hash: StarkHash) -> Result<Option<DHashT>, DbError>
 where
