+++ crates/client/rpc/src/methods/trace/trace_transaction.rs
@@ -18,7 +18,7 @@ use super::utils::{
     get_previous_block_substrate_hash, map_transaction_to_user_transaction, tx_execution_infos_to_tx_trace,
 };
 use crate::errors::StarknetRpcApiError;
-use crate::utils::get_block_by_block_hash;
+use crate::get_block_by_block_hash;
 use crate::Starknet;
 
 pub async fn trace_transaction<A, BE, G, C, P, H>(
